Deleted product are deleted from cloud also, if cloud deletion fails, deleted item gets save to deletions array in storage

+++ frontend/src/app/home/home.page.ts
@@ -203,7 +203,18 @@ export class HomePage implements OnInit {
     }, 5000);
   }
 
-  deleteItem(deletedItem: ILocalProduct) {
-    this.storageService.removeProduct(deletedItem.itemId);
+  async deleteItem(deletedItem: ILocalProduct) {
+    const deleteResponse = await this.api.deleteProduct(deletedItem.itemId);
+    if (deleteResponse.success) {
+      await this.storageService.removeProduct(deletedItem.itemId);
+      return;
+    }
+    await this.storageService.removeProduct(deletedItem.itemId);
+    await this.storageService.addDeletion({
+      itemId: deletedItem.itemId,
+      operation: 'DELETE',
+      clientUpdatedAt: new Date().toISOString(),
+    });
+    return;
   }
 }
